clear_unused_file
